+++ src/CoreEx.HealthChecks/Checks/TypedHttpClientHealthCheck.cs
@@ -1,3 +1,5 @@
+// Copyright (c) Avanade. Licensed under the MIT License. See https://github.com/Avanade/CoreEx
+
 using System;
 using System.Threading;
 using System.Threading.Tasks;
@@ -6,33 +8,28 @@ using Microsoft.Extensions.Diagnostics.HealthChecks;
 
 namespace CoreEx.Healthchecks.Checks
 {
-
-    /// <summary> Health check for Typed Http clients inheriting from <see cref="TypedHttpClientBase{T}"/>. </summary>
-    public class TypedHttpClientHealthCheck<T> : IHealthCheck
-        where T : TypedHttpClientBase<T>
+    /// <summary>
+    /// Health check for Typed Http clients inheriting from <see cref="TypedHttpClientBase{T}"/>.
+    /// </summary>
+    public class TypedHttpClientHealthCheck<T> : IHealthCheck where T : TypedHttpClientBase<T>
     {
         private readonly T _client;
 
-        /// <summary> constructor. </summary>
-        public TypedHttpClientHealthCheck(T client)
-        {
-            _client = client;
-        }
+        /// <summary>
+        /// Initializes a new instance of the <see cref="TypedHttpClientHealthCheck{T}"/> class.
+        /// </summary>
+        public TypedHttpClientHealthCheck(T client) => _client = client;
 
         /// <inheritdoc/>
         public async Task<HealthCheckResult> CheckHealthAsync(HealthCheckContext context, CancellationToken cancellationToken = default)
         {
             if (_client == null)
-            {
                 return HealthCheckResult.Unhealthy($"Typed Http client dependency for '{typeof(T)}' not resolved");
-            }
 
             try
             {
                 var result = await _client.HealthCheckAsync();
-
                 result.Response.EnsureSuccessStatusCode();
-
                 return HealthCheckResult.Healthy();
             }
             catch (Exception ex)