+++ src/CoreEx/Hosting/FileLockSynchronizer.cs
@@ -4,6 +4,7 @@ using CoreEx.Configuration;
 using System;
 using System.Collections.Concurrent;
 using System.IO;
+using System.Linq;
 
 namespace CoreEx.Hosting
 {